requirements-gui.txt
PyYAML>=6.0
streamlit